+++ lib/stage_craft.dart
@@ -1,4 +1,5 @@
 export 'src/field_configurators/field_configurator.dart';
 export 'src/stage_controller.dart';
 export 'src/widget_stage_data.dart';
+export 'src/widgets/color_picker.dart';
 export 'src/widgets/stage_craft.dart';
